feat: transaction lock system
