+++ packages/api/src/common/util/util.ts
@@ -1,7 +1,7 @@
 import { BadRequestException, NotFoundException } from "@nestjs/common";
 import { fromZonedTime, toZonedTime } from "date-fns-tz";
 
-import { IdType, MEntity } from "../model/entity.model";
+import { IdType, IEntity, MEntity } from "../base/entity.model";
 import { DB_TIMEZONE } from "./decorators/time-decorator";
 
 export const isEmptyObject = obj =>
@@ -87,9 +87,15 @@ export const makeObjectPropsFromDBTimezone = <T extends object | unknown>(
   }, {} as T);
 };
 
-type ModelInstance<Id extends IdType> = InstanceType<typeof MEntity<Id>>;
+type ModelInstance<Id extends IdType> = InstanceType<
+  typeof MEntity<IEntity<Id>, Id>
+>;
 
-type ModelClass<T extends MEntity<Id>, Id extends IdType, D = {}> = {
+type ModelClass<
+  T extends MEntity<IEntity<Id>, Id>,
+  Id extends IdType,
+  D = {},
+> = {
   prototype: T;
   modelName: string;
   new (data: D): T;
@@ -101,7 +107,7 @@ type ModelClass<T extends MEntity<Id>, Id extends IdType, D = {}> = {
  * @returns 표시할 모델 이름
  */
 const getModelName = <Id extends IdType, D = {}>(
-  name?: string | ModelClass<MEntity<Id>, Id, D>,
+  name?: string | ModelClass<MEntity<IEntity<Id>, Id>, Id, D>,
   model?: ModelInstance<Id>[] | null,
 ): string => {
   // 직접 지정된 이름이 있으면 사용
@@ -150,9 +156,11 @@ export const takeOne = <M>(values: M[]): M | null => {
  * @throws 배열이 비어있으면 NotFoundException 던짐
  * @throws 배열의 요소가 하나가 아니면 BadRequestException 던짐
  */
-export function takeOnlyOne<M extends MEntity<Id>, Id extends IdType, D = {}>(
-  name?: string | ModelClass<M, Id, D>,
-): (array: M[]) => M {
+export function takeOnlyOne<
+  M extends MEntity<IEntity<Id>, Id>,
+  Id extends IdType,
+  D = {},
+>(name?: string | ModelClass<M, Id, D>): (array: M[]) => M {
   return (array: M[]): M => {
     // 배열의 요소가 하나만 나왔는 지를 검증하는 함수
     // 배열의 요소가 하나가 아니면 예외 던짐
@@ -180,10 +188,11 @@ export function getUniqueArray<
  * @description 중복을 제외하고, 넣은 id가 모두 값이 잘 나왔는지를 체크해서 값을 얻는 함수
  * @description fetchAll에서 사용
  */
-export function takeAll<M extends MEntity<Id>, Id extends IdType, D = {}>(
-  ids: Id[],
-  name?: string | ModelClass<M, Id, D>,
-): (array: M[]) => M[] {
+export function takeAll<
+  M extends MEntity<IEntity<Id>, Id>,
+  Id extends IdType,
+  D = {},
+>(ids: Id[], name?: string | ModelClass<M, Id, D>): (array: M[]) => M[] {
   return (array: M[]): M[] => {
     const uniqueIds = getUniqueArray(ids);
     if (ids.some(id => !uniqueIds.includes(id))) {
@@ -198,9 +207,11 @@ export function takeAll<M extends MEntity<Id>, Id extends IdType, D = {}>(
 /**
  * @description 모델 배열이 비어있는 지 확인하고, 비어있으면 예외를 던지고, 비어있지 않으면 배열을 반환하는 함수
  */
-export function takeExist<M extends MEntity<Id>, Id extends IdType, D = {}>(
-  name?: string | ModelClass<M, Id, D>,
-): (array: M[]) => M[] {
+export function takeExist<
+  M extends MEntity<IEntity<Id>, Id>,
+  Id extends IdType,
+  D = {},
+>(name?: string | ModelClass<M, Id, D>): (array: M[]) => M[] {
   return (array: M[]) => {
     if (array.length === 0)
       throw new NotFoundException(`${getModelName(name, array)} is empty`);
@@ -232,3 +243,17 @@ export const takeToArray = <T>(data: T | T[]): T[] => {
   }
   return [data];
 };
+
+/**
+ * @description 배열의 요소를 순서대로 실행하는 함수
+ * @description async 함수를 순서대로 실행하는 것이 중요하다면 사용
+ */
+export async function runSequentially<T>(
+  items: T[],
+  runner: (item: T, index?: number) => Promise<void>,
+): Promise<void> {
+  await items.reduce(async (prev, item, idx) => {
+    await prev;
+    return runner(item, idx);
+  }, Promise.resolve());
+}
