+++ packages/api/src/feature/division/model/division.model.ts
@@ -2,7 +2,7 @@ import { InferSelectModel } from "drizzle-orm";
 
 import { IDivision } from "@clubs/interface/api/division/type/division.type";
 
-import { MEntity } from "@sparcs-clubs/api/common/model/entity.model";
+import { MEntity } from "@sparcs-clubs/api/common/base/entity.model";
 import { Division } from "@sparcs-clubs/api/drizzle/schema/division.schema";
 
 export type DivisionDBResult = InferSelectModel<typeof Division>;
