+++ packages/api/src/feature/semester/repository/activity.duration.repository.ts
@@ -1,7 +1,7 @@
 import { Injectable } from "@nestjs/common";
 import { gte, lte, not, or, SQL } from "drizzle-orm";
 
-import { BaseRepository } from "@sparcs-clubs/api/common/repository/base.repository";
+import { BaseRepository } from "@sparcs-clubs/api/common/base/base.repository";
 import { ActivityD } from "@sparcs-clubs/api/drizzle/schema/semester.schema";
 
 import {
